api/requirements.txt
flask
pika
mysql.connector
